+++ backend/src/main/java/com/reviewping/coflo/global/client/gitlab/response/ProjectInfoContent.java
@@ -3,13 +3,10 @@ package com.reviewping.coflo.global.client.gitlab.response;
 import java.util.Map;
 
 public record ProjectInfoContent(
-        Long commitCount,
-        Long branchCount,
-        Long mergerRequestCount,
-        Map<String, Double> languages) {
+        int commitCount, Long branchCount, Long mergerRequestCount, Map<String, Double> languages) {
 
     public static ProjectInfoContent of(
-            Long commitCount,
+            int commitCount,
             Long branchCount,
             Long mergerRequestCount,
             Map<String, Double> languages) {
